Move parsing logic to dedicated class.
